projects/weather-integration/python/requirements.txt
requests
python-dotenv
pyyaml
pyjwt
cryptography